+++ eedc/frontend/src/components/waermepumpe/index.ts
@@ -1,3 +1,3 @@
 export { WaermepumpeVergleich } from './WaermepumpeVergleich'
 export { WaermepumpeMonatsverlauf, WaermepumpeKostenvergleich, WaermepumpeMonatsTabelle, wpHatVergleich } from './WaermepumpeCharts'
-export { WaermepumpeModusSplit, hatModusSplit, type ModusSplitDaten } from './WaermepumpeModusSplit'
+export { WaermepumpeModusSplit, ModusSplitErklaerung, hatModusSplit, type ModusSplitDaten } from './WaermepumpeModusSplit'